Adjust logic and move it to useEditor
