+++ app/assets/js/components/Forms/RichTextArea.tsx
@@ -7,7 +7,7 @@ import { useValidation } from "./validations/hook";
 
 import classNames from "classnames";
 import { validateRichContentPresence } from "./validations/richContentPresence";
-import { isContentEmpty, useEditor, Editor as RichTextEditor, RichContent } from "turboui";
+import { useEditor, Editor as RichTextEditor, RichContent } from "turboui";
 import { useRichEditorHandlers } from "@/hooks/useRichEditorHandlers";
 
 interface RichTextAreaProps {
@@ -27,7 +27,6 @@ interface RichTextAreaProps {
   showToolbarTopBorder?: boolean;
   readonly?: boolean;
   hideToolbar?: boolean;
-  autosaveKey?: string;
 }
 
 const DEFAULT_VALUES = {
@@ -38,8 +37,6 @@ const DEFAULT_VALUES = {
   fontWeight: "font-medium",
   height: "min-h-[250px]",
 };
-
-const AUTOSAVE_TTL_MS = 1000 * 60 * 60 * 6;
 
 export function RichTextArea(props: RichTextAreaProps) {
   props = { ...DEFAULT_VALUES, ...props };
@@ -71,10 +68,6 @@ function ReadonlyContent(props: RichTextAreaProps) {
 function Editor(props: RichTextAreaProps & { error: boolean }) {
   const form = useFormContext();
   const [value, setValue] = useFieldValue(props.field);
-  const autosaveKey = React.useMemo(() => {
-    if (props.autosaveKey) return props.autosaveKey;
-    return `operately:richtext:${window.location.pathname}${window.location.search}:${props.field}`;
-  }, [props.autosaveKey, props.field]);
 
   const handlers = useRichEditorHandlers({ scope: props.mentionSearchScope });
   const editor = useEditor({
@@ -86,18 +79,6 @@ function Editor(props: RichTextAreaProps & { error: boolean }) {
     },
     onUpdate: ({ json }) => {
       setValue(json);
-      if (autosaveKey) {
-        try {
-          if (isContentEmpty(json)) {
-            sessionStorage.removeItem(autosaveKey);
-            return;
-          }
-
-          sessionStorage.setItem(autosaveKey, JSON.stringify({ savedAt: Date.now(), content: json }));
-        } catch (error) {
-          console.warn("Failed to store rich text autosave state", error);
-        }
-      }
     },
     onUploadStatusChange: (status) => {
       if (status) {
@@ -113,27 +94,6 @@ function Editor(props: RichTextAreaProps & { error: boolean }) {
       editor.editor.commands.setContent(value);
     }
   }, [editor.editor]);
-
-  React.useEffect(() => {
-    if (!autosaveKey) return;
-    if (!isContentEmpty(value)) return;
-
-    try {
-      const stored = sessionStorage.getItem(autosaveKey);
-      if (!stored) return;
-
-      const parsed = JSON.parse(stored) as { savedAt?: number; content?: unknown };
-      if (!parsed?.content) return;
-      if (parsed.savedAt && Date.now() - parsed.savedAt > AUTOSAVE_TTL_MS) {
-        sessionStorage.removeItem(autosaveKey);
-        return;
-      }
-
-      setValue(parsed.content as any);
-    } catch (error) {
-      console.warn("Failed to restore rich text autosave state", error);
-    }
-  }, [autosaveKey, setValue, value]);
 
   return (
     <div className={wrapperClassName(props)}>
